FIX: CSP issue, anonymous webinar visibility
The Zoom SDK endpoint requires a modification to the CSP, Zoom recently
started using a service worker via `blob:`, so that needs to be allowed.

I also noticed that the whole webinar block is hidden to anonymous
users, it makes sense to show that section, even though users need
accounts to register/join a webinar event.

Another fix included here: we need to pass something to the Zoom SDK for
"username", so if names are empty, usernames are used.